feat(stenography): Support the Plover HID Protocol
This commit adds tentative support for the Plover HID Protocol, a
simple HID-based protocol that works over both USB and BLE.
The main value-add of this feature is that it allows ZMK to be used for
interfacing with Plover from the Open Steno Project without Plover
having to intercept all keyboard keypresses.

Usually hobbyist steno machines use legacy serial port protocols for
this, but since we can't use those over BLE we might as well develop a
protocol that works both for USB and BLE and at the same time remove
some of the pain points of those old protocols.